HW3+PT3

+++ src/com/softserve/edu03/pt/task1/Calculation.java
@@ -1,12 +1,15 @@
 package com.softserve.edu03.pt.task1;
+import com.softserve.edu03.pt.task1.Calculation;
 import java.util.Scanner;
+
+
 public class Calculation {
 
     public double Summary(double a, double b) {
-        return (double) (a + b);
+        return  (a + b);
     }
     public double Average(double a, double b) {
-        return (double) ((a + b) / 2.0);
+        return ((a + b) / 2.0);
     }
 
     public static int getNumber(String prompt) {
